Add rock paper scissors functionality
Added the functionality to play rock paper scissors with the terminal
in and the prompt. Press CTRL+SHIFT+J to see the outputs on the server
console. getHumanChoice is used for getting the user input.
getComputerChoice is for getting the randomised computer input.
playRound is for evaluating the outcome of a rps round. playGame
is for playing the game for 5 rounds. determineWinner is for
determining which player has more points and the player with the most
points is going to be the winner. printScore prints the score of the
player and the computer in the console (web) / terminal (node).