Add event listener rps buttons
Furthermore, added comment to playRound function for more explainability
about the return values of the function as it does not speak for itself.

The event listener is added so instead of the human player having to
type out what hand they want to play, they can interact with buttons
to choose which hand he chooses.